+++ src/main/java/edu/ie3/datamodel/io/source/TypeSource.java
@@ -29,6 +29,8 @@ import java.util.UUID;
  * @since 08.04.20
  */
 public class TypeSource extends EntitySource {
+  private static final String SUB_DIRECTORY = "/type";
+
   // factories
   private final OperatorInputFactory operatorInputFactory;
   private final Transformer2WTypeInputFactory transformer2WTypeInputFactory;
@@ -76,7 +78,44 @@ public class TypeSource extends EntitySource {
    * @return a map of UUID to object- and uuid-unique {@link Transformer2WTypeInput} entities
    */
   public Map<UUID, Transformer2WTypeInput> getTransformer2WTypes() throws SourceException {
-    return getEntities(Transformer2WTypeInput.class, dataSource, transformer2WTypeInputFactory);
+    return getTransformer2WTypes(true);
+  }
+
+  /**
+   * Returns a set of build in {@link Transformer2WTypeInput} instances within a map by UUID.
+   *
+   * @return a map of UUID to object- and uuid-unique {@link Transformer2WTypeInput} entities
+   */
+  public static Map<UUID, Transformer2WTypeInput> getStandardTransformer2WTypes()
+      throws SourceException {
+    return new TypeSource(getBuildInSource(Transformer2WTypeInput.class, SUB_DIRECTORY))
+        .getTransformer2WTypes(false);
+  }
+
+  /**
+   * Returns a set of {@link Transformer2WTypeInput} instances within a map by UUID.
+   *
+   * <p>This set has to be unique in the sense of object uniqueness but also in the sense of {@link
+   * UUID} uniqueness of the provided {@link Transformer2WTypeInput} which has to be checked
+   * manually, as {@link Transformer2WTypeInput#equals(Object)} is NOT restricted on the uuid of
+   * {@link Transformer2WTypeInput}.
+   *
+   * @param withBuildIn if true the standard transformer types will be included if their uuid is not
+   *     overwritten by the source
+   * @return a map of UUID to object- and uuid-unique {@link Transformer2WTypeInput} entities
+   */
+  private Map<UUID, Transformer2WTypeInput> getTransformer2WTypes(boolean withBuildIn)
+      throws SourceException {
+    Map<UUID, Transformer2WTypeInput> types =
+        getEntities(Transformer2WTypeInput.class, dataSource, transformer2WTypeInputFactory);
+
+    if (withBuildIn) {
+      Map<UUID, Transformer2WTypeInput> allTypes = getStandardTransformer2WTypes();
+      allTypes.putAll(types);
+      return allTypes;
+    }
+
+    return types;
   }
 
   /**
@@ -102,7 +141,40 @@ public class TypeSource extends EntitySource {
    * @return a map of UUID to object- and uuid-unique {@link LineTypeInput} entities
    */
   public Map<UUID, LineTypeInput> getLineTypes() throws SourceException {
-    return getEntities(LineTypeInput.class, dataSource, lineTypeInputFactory);
+    return getLineTypes(true);
+  }
+
+  /**
+   * Returns a set of build in {@link LineTypeInput} instances within a map by UUID.
+   *
+   * @return a map of UUID to object- and uuid-unique {@link LineTypeInput} entities
+   */
+  public static Map<UUID, LineTypeInput> getStandardLineTypes() throws SourceException {
+    return new TypeSource(getBuildInSource(LineTypeInput.class, SUB_DIRECTORY)).getLineTypes(false);
+  }
+
+  /**
+   * Returns a set of {@link LineTypeInput} instances within a map by UUID.
+   *
+   * <p>This set has to be unique in the sense of object uniqueness but also in the sense of {@link
+   * UUID} uniqueness of the provided {@link LineTypeInput} which has to be checked manually, as
+   * {@link LineTypeInput#equals(Object)} is NOT restricted on the uuid of {@link LineTypeInput}.
+   *
+   * @param withBuildIn if true the standard line types will be included if their uuid is not
+   *     overwritten by the source
+   * @return a map of UUID to object- and uuid-unique {@link LineTypeInput} entities
+   */
+  private Map<UUID, LineTypeInput> getLineTypes(boolean withBuildIn) throws SourceException {
+    Map<UUID, LineTypeInput> types =
+        getEntities(LineTypeInput.class, dataSource, lineTypeInputFactory);
+
+    if (withBuildIn) {
+      Map<UUID, LineTypeInput> allTypes = getStandardLineTypes();
+      allTypes.putAll(types);
+      return allTypes;
+    }
+
+    return types;
   }
 
   /**
@@ -116,7 +188,44 @@ public class TypeSource extends EntitySource {
    * @return a map of UUID to object- and uuid-unique {@link Transformer3WTypeInput} entities
    */
   public Map<UUID, Transformer3WTypeInput> getTransformer3WTypes() throws SourceException {
-    return getEntities(Transformer3WTypeInput.class, dataSource, transformer3WTypeInputFactory);
+    return getTransformer3WTypes(true);
+  }
+
+  /**
+   * Returns a set of build in {@link Transformer3WTypeInput} instances within a map by UUID.
+   *
+   * @return a map of UUID to object- and uuid-unique {@link Transformer3WTypeInput} entities
+   */
+  public static Map<UUID, Transformer3WTypeInput> getStandardTransformer3WTypes()
+      throws SourceException {
+    return new TypeSource(getBuildInSource(Transformer3WTypeInput.class, SUB_DIRECTORY))
+        .getTransformer3WTypes(false);
+  }
+
+  /**
+   * Returns a set of {@link Transformer3WTypeInput} instances within a map by UUID.
+   *
+   * <p>This set has to be unique in the sense of object uniqueness but also in the sense of {@link
+   * UUID} uniqueness of the provided {@link Transformer3WTypeInput} which has to be checked
+   * manually, as {@link Transformer3WTypeInput#equals(Object)} is NOT restricted on the uuid of
+   * {@link Transformer3WTypeInput}.
+   *
+   * @param withBuildIn if true the standard transformer types will be included if their uuid is not
+   *     overwritten by the source
+   * @return a map of UUID to object- and uuid-unique {@link Transformer3WTypeInput} entities
+   */
+  private Map<UUID, Transformer3WTypeInput> getTransformer3WTypes(boolean withBuildIn)
+      throws SourceException {
+    Map<UUID, Transformer3WTypeInput> types =
+        getEntities(Transformer3WTypeInput.class, dataSource, transformer3WTypeInputFactory);
+
+    if (withBuildIn) {
+      Map<UUID, Transformer3WTypeInput> allTypes = getStandardTransformer3WTypes();
+      allTypes.putAll(types);
+      return allTypes;
+    }
+
+    return types;
   }
 
   /**